Need fix the last 4 sort algo

+++ SortAlg.h
@@ -1,30 +1,30 @@
 #pragma once
 
-void selectionSort(int* array, int size, int &count_assign, int& count_compare, float& Time);
+void selectionSort(int* array, int size, long long&count_assign, long long& count_compare, float& Time);
 
-void insertionSort(int* array, int size, int &count_assign, int& count_compare, float& Time);
+void insertionSort(int* array, int size, long long&count_assign, long long& count_compare, float& Time);
 
-void bubbleSort(int* array, int size, int &count_assign, int& count_compare, float& Time);
+void bubbleSort(int* array, int size, long long&count_assign, long long& count_compare, float& Time);
 
-void heapSort(int* array, int size, int &count_assign, int& count_compare, float& Time);
+void heapSort(int* array, int size, long long&count_assign, long long& count_compare, float& Time);
 
-void countingSort(int* array, int size, int &count_assign, int& count_compare, float& Time);
+void countingSort(int* array, int size, long long&count_assign, long long& count_compare, float& Time);
 
-void radixSort(int* array, int size, int &count_assign, int& count_compare, float& Time);
+void radixSort(int* array, int size, long long&count_assign, long long& count_compare, float& Time);
 
-void qSortCount(int* array, int size, int &count_assign, int& count_compare, float& Time);
+void qSortCount(int* array, int size, long long&count_assign, long long& count_compare, float& Time);
 
-void mergeSortCount(int* array, int size, int &count_assign, int& count_compare, float& Time);
+void mergeSortCount(int* array, int size, long long&count_assign, long long& count_compare, float& Time);
 
-void quickSort(int* array, int size, int &count_assign, int& count_compare);
+void quickSort(int* array, int size, long long&count_assign, long long& count_compare);
 
-void mergeSort(int* array, int size, int &count_assign, int& count_compare);
+void mergeSort(int* array, int size, long long&count_assign, long long& count_compare);
 
-void radixSort(int* array, int size, int &count_assign, int& count_compare, float& Time);
+void radixSort(int* array, int size, long long&count_assign, long long& count_compare, float& Time);
 
-void shellSort(int* array, int size, int &count_assign, int& count_compare, float& Time);
+void shellSort(int* array, int size, long long&count_assign, long long& count_compare, float& Time);
 
-void shakerSort(int* array, int size, int &count_assign, int& count_compare, float& Time);
+void shakerSort(int* array, int size, long long&count_assign, long long& count_compare, float& Time);
 
-void SortChoosing(int* array, int size, int sortType, int &count_assign, int &count_compare, float &Time);
+void SortChoosing(int* array, int size, int sortType, long long& count_assign, long long& count_compare, float &Time);
